Version 1.3 can be able to get all needed data
In this version, I have added remaining functions that can be able to fetch others needed data from GitHub, but there is still a problem that GitHub limits the number of returned results (30). For this reason, GitHub terminated my request when I was fetching too much data. I struggling with how to get all needed data from GitHub even if it will take a long time.